+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Added
 - user: organizations: tooltip to show org name
 - user: CalendarLine (chart): tooltip to show grouped contributions per week
+- components: PolarAreaChart: tooltip to show percentage
 
 (feature/area-chart-update)
 #### Fixed